Update WasmCredentialValidator to use new resolver types
