+++ bindings/wasm/docs/api-reference.md
@@ -200,6 +200,11 @@ See <code>IVerifierOptions</code>.</p>
 <dl>
 <dt><a href="#StateMetadataEncoding">StateMetadataEncoding</a></dt>
 <dd></dd>
+<dt><a href="#MethodRelationship">MethodRelationship</a></dt>
+<dd></dd>
+<dt><a href="#DIDType">DIDType</a></dt>
+<dd><p>Supported types representing a DID that can be generated by the storage interface.</p>
+</dd>
 <dt><a href="#StatusCheck">StatusCheck</a></dt>
 <dd><p>Controls validation behaviour when checking whether or not a credential has been revoked by its
 <a href="https://www.w3.org/TR/vc-data-model/#status"><code>credentialStatus</code></a>.</p>
@@ -241,11 +246,6 @@ This variant is the default used if no other variant is specified when construct
 </dd>
 <dt><a href="#FirstError">FirstError</a></dt>
 <dd><p>Return after the first error occurs.</p>
-</dd>
-<dt><a href="#MethodRelationship">MethodRelationship</a></dt>
-<dd></dd>
-<dt><a href="#DIDType">DIDType</a></dt>
-<dd><p>Supported types representing a DID that can be generated by the storage interface.</p>
 </dd>
 <dt><a href="#KeyType">KeyType</a></dt>
 <dd></dd>
@@ -2144,10 +2144,10 @@ Deserializes an instance from a JSON object.
     * [.checkStructure(credential)](#CredentialValidator.checkStructure)
     * [.checkExpiresOnOrAfter(credential, timestamp)](#CredentialValidator.checkExpiresOnOrAfter)
     * [.checkIssuedOnOrBefore(credential, timestamp)](#CredentialValidator.checkIssuedOnOrBefore)
-    * [.verifySignature(credential, trusted_issuers, options)](#CredentialValidator.verifySignature)
+    * [.verifySignature(credential, trustedIssuers, options)](#CredentialValidator.verifySignature)
     * [.check_subject_holder_relationship(credential, holder_url, relationship)](#CredentialValidator.check_subject_holder_relationship)
     * [.checkStatus(credential, trustedIssuers, statusCheck)](#CredentialValidator.checkStatus)
-    * [.extractIssuer(credential)](#CredentialValidator.extractIssuer) ⇒ [<code>IotaDID</code>](#IotaDID)
+    * [.extractIssuer(credential)](#CredentialValidator.extractIssuer) ⇒ [<code>CoreDID</code>](#CoreDID) \| [<code>StardustDID</code>](#StardustDID)
 
 <a name="CredentialValidator.validate"></a>
 
@@ -2182,7 +2182,7 @@ An error is returned whenever a validated condition is not satisfied.
 | Param | Type |
 | --- | --- |
 | credential | [<code>Credential</code>](#Credential) | 
-| issuer | [<code>Document</code>](#Document) \| [<code>ResolvedDocument</code>](#ResolvedDocument) | 
+| issuer | [<code>StardustDocument</code>](#StardustDocument) \| [<code>CoreDocument</code>](#CoreDocument) | 
 | options | [<code>CredentialValidationOptions</code>](#CredentialValidationOptions) | 
 | fail_fast | <code>number</code> | 
 
@@ -2226,7 +2226,7 @@ Validate that the credential is issued on or before the specified timestamp.
 
 <a name="CredentialValidator.verifySignature"></a>
 
-### CredentialValidator.verifySignature(credential, trusted_issuers, options)
+### CredentialValidator.verifySignature(credential, trustedIssuers, options)
 Verify the signature using the DID Document of a trusted issuer.
 
 # Warning
@@ -2241,7 +2241,7 @@ to verify the credential's signature will be made and an error is returned upon 
 | Param | Type |
 | --- | --- |
 | credential | [<code>Credential</code>](#Credential) | 
-| trusted_issuers | <code>Array.&lt;(Document\|ResolvedDocument)&gt;</code> | 
+| trustedIssuers | <code>Array.&lt;(StardustDocument\|CoreDocument)&gt;</code> | 
 | options | [<code>VerifierOptions</code>](#VerifierOptions) | 
 
 <a name="CredentialValidator.check_subject_holder_relationship"></a>
@@ -2270,12 +2270,12 @@ Only supports `BitmapRevocation2022`.
 | Param | Type |
 | --- | --- |
 | credential | [<code>Credential</code>](#Credential) | 
-| trustedIssuers | <code>Array.&lt;(Document\|ResolvedDocument)&gt;</code> | 
+| trustedIssuers | <code>Array.&lt;(StardustDocument\|CoreDocument)&gt;</code> | 
 | statusCheck | <code>number</code> | 
 
 <a name="CredentialValidator.extractIssuer"></a>
 
-### CredentialValidator.extractIssuer(credential) ⇒ [<code>IotaDID</code>](#IotaDID)
+### CredentialValidator.extractIssuer(credential) ⇒ [<code>CoreDID</code>](#CoreDID) \| [<code>StardustDID</code>](#StardustDID)
 Utility for extracting the issuer field of a `Credential` as a DID.
 
 ### Errors
@@ -7329,6 +7329,16 @@ This is possible because Ed25519 is birationally equivalent to Curve25519 used b
 
 ## StateMetadataEncoding
 **Kind**: global variable  
+<a name="MethodRelationship"></a>
+
+## MethodRelationship
+**Kind**: global variable  
+<a name="DIDType"></a>
+
+## DIDType
+Supported types representing a DID that can be generated by the storage interface.
+
+**Kind**: global variable  
 <a name="StatusCheck"></a>
 
 ## StatusCheck
@@ -7405,16 +7415,6 @@ Return all errors that occur during validation.
 
 ## FirstError
 Return after the first error occurs.
-
-**Kind**: global variable  
-<a name="MethodRelationship"></a>
-
-## MethodRelationship
-**Kind**: global variable  
-<a name="DIDType"></a>
-
-## DIDType
-Supported types representing a DID that can be generated by the storage interface.
 
 **Kind**: global variable  
 <a name="KeyType"></a>
